requirements.txt
requests>=2.28
PyYAML>=6.0
python-dotenv>=1.0
rich>=13.0.0
ruff>=0.14.14